rm test dir temporary & add loading partly
